Move `ir_tree::Module` struct to its own file
